backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
deepface==0.0.79
tf-keras==2.15.0
pymongo==4.6.0
numpy==1.24.3
python-dotenv==1.0.0
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
